+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 * Moved setting the output path to outside of `PyTrivialOpenGL_TestsAndExamples` module.
 * Improved `ExampleManager` and `ExampleRunner` modules from `PyTrivialOpenGL_TestsAndExamples`.
 * Added `PyTrivialOpenGL` package information to the package.
+* Changed default behaviour of `Window.center` method. Now parameters width, height and size are interpreted by default as draw area width, height and size.
+